requirements_freeze.txt
six==1.12.0
click==5.1
pathtools==0.1.2
watchdog==0.9.0
libsass==0.18.0
pyaml==19.4.1
PyYAML==5.1
colorama==0.4.1
colorlog==4.0.2
